requirements.txt
# requirements.txt
streamlit
pandas
psycopg2-binary
python-dotenv